+++ arch/x86/core/intel64/irq.c
@@ -13,19 +13,11 @@
 #include <zephyr/irq.h>
 #include <zephyr/logging/log.h>
 #include <zephyr/sys/iterable_sections.h>
-#include <x86_mmu.h>
-#include <zephyr/init.h>
+
 
 LOG_MODULE_DECLARE(os, CONFIG_KERNEL_LOG_LEVEL);
 
 unsigned char _irq_to_interrupt_vector[CONFIG_MAX_IRQ_LINES];
-
-/*
- * The low-level interrupt code consults these arrays to dispatch IRQs, so
- * so be sure to keep locore.S up to date with any changes. Note the indices:
- * use (vector - IV_IRQS), since exception vectors do not appear here.
- */
-
 #define NR_IRQ_VECTORS (IV_NR_VECTORS - IV_IRQS)  /* # vectors free for IRQs */
 
 void (*x86_irq_funcs[NR_IRQ_VECTORS])(const void *arg);
@@ -139,48 +131,6 @@ int arch_irq_connect_dynamic(unsigned int irq, unsigned int priority,
 	return vector;
 }
 
-#ifdef CONFIG_IRQ_OFFLOAD
-#include <zephyr/irq_offload.h>
-
-void arch_irq_offload(irq_offload_routine_t routine, const void *parameter)
-{
-	x86_irq_funcs[CONFIG_IRQ_OFFLOAD_VECTOR - IV_IRQS] = routine;
-	x86_irq_args[CONFIG_IRQ_OFFLOAD_VECTOR - IV_IRQS] = parameter;
-	__asm__ volatile("int %0" : : "i" (CONFIG_IRQ_OFFLOAD_VECTOR)
-			  : "memory");
-	x86_irq_funcs[CONFIG_IRQ_OFFLOAD_VECTOR - IV_IRQS] = NULL;
-}
-
-#endif /* CONFIG_IRQ_OFFLOAD */
-
-#if defined(CONFIG_SMP)
-
-int arch_smp_init(void)
-{
-	/*
-	 * z_sched_ipi() doesn't have the same signature as a typical ISR, so
-	 * we fudge it with a cast. the argument is ignored, no harm done.
-	 */
-
-	x86_irq_funcs[CONFIG_SCHED_IPI_VECTOR - IV_IRQS] =
-		(void *) z_sched_ipi;
-
-	/* TLB shootdown handling */
-	x86_irq_funcs[CONFIG_TLB_IPI_VECTOR - IV_IRQS] = z_x86_tlb_ipi;
-	return 0;
-}
-
-SYS_INIT(arch_smp_init, PRE_KERNEL_1, CONFIG_KERNEL_INIT_PRIORITY_DEFAULT);
-
-/*
- * it is not clear exactly how/where/why to abstract this, as it
- * assumes the use of a local APIC (but there's no other mechanism).
- */
-void arch_sched_ipi(void)
-{
-	z_loapic_ipi(0, LOAPIC_ICR_IPI_OTHERS, CONFIG_SCHED_IPI_VECTOR);
-}
-#endif
 
 /* The first bit is used to indicate whether the list of reserved interrupts
  * have been initialized based on content stored in the irq_alloc linker
